refactor: added init command as aver cli command
