+++ packages/@averjs/typescript-runtime/lib/index.ts
@@ -1,10 +1,7 @@
-import { resolve } from 'path';
-import { existsSync, writeFileSync, readFileSync } from 'fs';
-import { copy } from 'fs-extra';
 import { register } from 'ts-node';
 import parseArgs, { ParsedArgs } from 'minimist';
-import ora from 'ora';
 import AverCli from '@averjs/cli';
+import TypescriptInitCommand from './ts-init';
 
 interface AvailableCommands extends ParsedArgs {
   init?: boolean
@@ -12,46 +9,16 @@ interface AvailableCommands extends ParsedArgs {
 
 export default class CLI {
   argv: AvailableCommands = parseArgs(process.argv.slice(2));
-
-  get tsConfigPath() { return resolve(process.env.PROJECT_PATH, '../tsconfig.json'); }
-  get tsConfigServerPath() { return resolve(process.env.PROJECT_PATH, '../tsconfig.server.json'); }
-  get eslintrcPath() { return resolve(process.env.PROJECT_PATH, '../.eslintrc.js'); }
   
   async run() {
-    if(this.argv.init) {
-      await this.init();
-      return;
-    }
+    const typescriptInitCommand = new TypescriptInitCommand();
 
     register({
-      project: this.tsConfigServerPath
+      project: typescriptInitCommand.tsConfigServerPath
     });
 
     const avercli = new AverCli();
+    avercli.addCommand(typescriptInitCommand);
     avercli.run();
-  }
-
-  async init() {
-    const tsConfigSpinner = ora('Copying tsconfig.json').start();
-    if (!existsSync(this.tsConfigPath)) {
-      await copy(
-        resolve(__dirname, '../templates/tsconfig.json'),
-        this.tsConfigPath
-      );
-      tsConfigSpinner.succeed();
-    } else {
-      tsConfigSpinner.info('Skipping because tsconfig.json already exists.');
-    }
-
-    const tsConfigServerSpinner = ora('Copying tsconfig.server.json').start();
-    if (!existsSync(this.tsConfigServerPath)) {
-      await copy(
-        resolve(__dirname, '../templates/tsconfig.server.json'),
-        this.tsConfigServerPath
-      );
-      tsConfigServerSpinner.succeed();
-    } else {
-      tsConfigServerSpinner.info('Skipping because tsconfig.server.json already exists.');
-    }
   }
 }
